[Mod] move OffsetConverter module to vnpy.trader
