+++ vnpy/trader/converter.py
